Refactor email template handling
